edit data type

+++ models/Teacher.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, dataTypes) => {
     "Teacher",
     {
       teacherId: {
-        type: dataTypes.INTEGER(4),
+        type: dataTypes.INTEGER,
         primaryKey: true,
         allowNull: false,
         autoIncrement: true,
